rename profile to application to resolve any ambiguity with listener profiles
