+++ .eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-plusplus': 'off',
+    'import/no-cycle': 'off',
   },
   parserOptions: {
     parser: 'babel-eslint',
